+++ mango3-studio/src/server_functions/mod.rs
@@ -1,5 +1,7 @@
+mod pages;
 mod posts;
 mod websites;
 
+pub use pages::*;
 pub use posts::*;
 pub use websites::*;