+++ mango3-websites/src/pages/mod.rs
@@ -1,5 +1,7 @@
 mod index_page;
+mod show_page_page;
 mod show_post_page;
 
 pub use index_page::IndexPage;
+pub use show_page_page::ShowPagePage;
 pub use show_post_page::ShowPostPage;